Added Colors and made changes to elevator bp
added comments for documentation, got rid of button pressing and replaced it with color changing for now. I also disabled the disabling feature for testing purposes, but due to the max height and story height you should be able to go up to a max of 10 times before the elevator will close the doors bu not move.